Refactor packaging context into modules
Split the large context.rs into
context/{axes,builder,files,manifest,prompts}.rs and re-export them via
a context mod. Also remove legacy token UI binding and related CSS, and
update the Access Control wording in the UI.
