+++ kapsl-runtime/ui/index.html
@@ -782,7 +782,7 @@
                         <div class="section-header access-section-header">
                             <div class="section-title">
                                 <h2>Access Control</h2>
-                                <p>Users, API keys, and legacy role tokens.</p>
+                                <p>Users, API keys, and scoped access.</p>
                             </div>
                             <button
                                 class="btn btn-secondary"
@@ -856,54 +856,6 @@
                                         >
                                     </div>
                                 </div>
-                            </section>
-
-                            <section class="access-panel">
-                                <h3>Legacy Role Tokens</h3>
-                                <p class="access-help">
-                                    Compatibility tokens for reader/writer/admin
-                                    roles.
-                                </p>
-                                <form id="legacy-tokens-form" class="stack">
-                                    <input
-                                        id="legacy-reader-token"
-                                        class="access-input"
-                                        type="password"
-                                        placeholder="reader token"
-                                    />
-                                    <input
-                                        id="legacy-writer-token"
-                                        class="access-input"
-                                        type="password"
-                                        placeholder="writer token"
-                                    />
-                                    <input
-                                        id="legacy-admin-token"
-                                        class="access-input"
-                                        type="password"
-                                        placeholder="admin token"
-                                    />
-                                    <div class="access-inline-form">
-                                        <button
-                                            class="btn btn-primary"
-                                            id="legacy-save-btn"
-                                            type="submit"
-                                        >
-                                            Save Tokens
-                                        </button>
-                                        <button
-                                            class="btn btn-secondary"
-                                            id="legacy-refresh-btn"
-                                            type="button"
-                                        >
-                                            Reload
-                                        </button>
-                                    </div>
-                                </form>
-                                <div
-                                    id="legacy-feedback"
-                                    class="access-status-line"
-                                ></div>
                             </section>
                         </div>
